[#739] Cleaned up vos for braodband
